requirements.txt
numpy<2
opencv-python
pillow
torch
torchvision
stable-baselines3[extra]
shimmy>=2.0
gym==0.21.0
gym-retro
tensorboard
ultralytics
